# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/poas/{poaid}/actividades": {
      "post": {
        "description": "",
        "operationId": "PoasController_agregarActividades",
        "parameters": [
          {
            "name": "poaid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearActividadesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Actividad agregada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrearActividadesResponseDto"
                }
              }
            }
          }
        },
        "summary": "Agregar actividades",
        "tags": [
          "POAs"
        ]
      }
    },
    "/api/v1/poas/{poaid}/presentar": {
      "post": {
        "description": "",
        "operationId": "PoasController_presentarPoa",
        "parameters": [
          {
            "name": "poaid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentarPoasDto"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Error al presentar la POA",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentarPoasResponseErrorDto"
                }
              }
            }
          }
        },
        "summary": "Presentar la poa para el envio",
        "tags": [
          "POAs"
        ]
      }
    },
    "/api/v1/poas/{poaid}/cancelar-envio": {
      "post": {
        "description": "",
        "operationId": "PoasController_cancelarEnvio",
        "parameters": [
          {
            "name": "poaid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Se regresa el estado de la POA de enviada a en progreso"
          }
        },
        "summary": "Cancela el envio de la POA",
        "tags": [
          "POAs"
        ]
      }
    },
    "/api/v1/catalogos/banco-actividades": {
      "get": {
        "description": "",
        "operationId": "CatalogoController_getBancoActividades",
        "parameters": [
          {
            "name": "Busqueda",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Obtiene el catálogo de actividades del banco de actividades",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BancoActividadesDataDto"
                }
              }
            }
          }
        },
        "summary": "Muestra el catalogo de actividades",
        "tags": [
          "Catalogos"
        ]
      }
    },
    "/api/v1/catalogos/banco-actividades/{id}": {
      "get": {
        "description": "",
        "operationId": "CatalogoController_getBancoActividadesById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtiene la actividad con el ID especificado dentro del banco",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BancoIdDto"
                }
              }
            }
          }
        },
        "summary": "Buscar actividad por ID",
        "tags": [
          "Catalogos"
        ]
      }
    },
    "/api/v1/catalogos/banco-actividades/{id}/sub-actividades-sugeridas": {
      "get": {
        "description": "",
        "operationId": "CatalogoController_getSubActividadesSugeridas",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtiene las sub-actividades sugeridas para una actividad específica",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActividadSugeridaDataDto"
                }
              }
            }
          }
        },
        "summary": "Entrega sub-actividades sugeridas",
        "tags": [
          "Catalogos"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Silverback <3",
    "description": "Especificación técnica de los endpoints para auditoría, gestión documental y reportes.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EquipoAuditorDto": {
        "type": "object",
        "properties": {
          "total_participantes": {
            "type": "number",
            "description": "Numero de participantes en la actividad",
            "example": 2
          },
          "auditores_ids": {
            "description": "Lista con los IDs de los auditores que participaran en la actividad",
            "example": [
              "uuid-auditor-1",
              "uuid-auditor-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "total_participantes",
          "auditores_ids"
        ]
      },
      "CrearActividadesDto": {
        "type": "object",
        "properties": {
          "banco_actividad_id": {
            "type": "string",
            "description": "ID de la actividad dentro del banco de actividades",
            "example": "banco-uuid-1"
          },
          "titulo": {
            "type": "string",
            "description": "Titulo de la actividad",
            "example": "Revisión al rubro de Obra Pública"
          },
          "justificacion": {
            "type": "string",
            "description": "Justificacion del porque de la Actividad",
            "example": "Verificar el adecuado cumplimiento en mi Centro Universitario..."
          },
          "objetivo_general": {
            "type": "string",
            "description": "Objetivo de la actividad",
            "example": "Asegurar la correcta aplicación..."
          },
          "objetivos_especificos": {
            "type": "string",
            "description": "Objetivos especificos de la actividad",
            "example": "Revisar el ejercicio de los recursos y los procesos.."
          },
          "metas": {
            "type": "string",
            "description": "Metas de la actividad",
            "example": "06 revisiones"
          },
          "indicadores": {
            "type": "string",
            "description": "Objetivos especificos de la actividad",
            "example": "Numero de revisiones realizadas entre numero de revisiones programadas"
          },
          "equipo_auditor": {
            "description": "Informacion sobre el equipo de participantes en esta actividad",
            "allOf": [
              {
                "$ref": "#/components/schemas/EquipoAuditorDto"
              }
            ]
          }
        },
        "required": [
          "titulo",
          "justificacion",
          "objetivo_general",
          "objetivos_especificos",
          "metas",
          "indicadores",
          "equipo_auditor"
        ]
      },
      "CrearActividadesResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la nueva actividad agregada",
            "example": "act-nueva-uuid-999"
          },
          "folio": {
            "type": "string",
            "description": "Folio de la nueva actividad agregada",
            "example": "03"
          },
          "mensaje": {
            "type": "string",
            "description": "Mensaje API de que la accion fue completada con exito",
            "example": "Actividad guardada exitosamente."
          }
        },
        "required": [
          "id",
          "folio",
          "mensaje"
        ]
      },
      "PresentarPoasDto": {
        "type": "object",
        "properties": {
          "poa_id": {
            "type": "string",
            "description": "ID de la POA a presentar",
            "example": "uuid-poa-1234"
          },
          "estado_anterior": {
            "type": "string",
            "description": "Estado anterior de la POA",
            "example": "EN_PROGRESO"
          },
          "estado_nuevo": {
            "type": "string",
            "description": "Estado nuevo de la POA",
            "example": "EN_REVISION"
          },
          "fecha_envio": {
            "type": "string",
            "description": "Fecha y hora en la que se envio la POA",
            "example": "2026-06-28T19:30:00Z"
          },
          "mensaje": {
            "type": "string",
            "description": "ID de la POA a presentar",
            "example": "uuid-poa-1234"
          }
        },
        "required": [
          "poa_id",
          "estado_anterior",
          "estado_nuevo",
          "fecha_envio",
          "mensaje"
        ]
      },
      "PresentarPoasResponseErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Titulo del error",
            "example": "Validación de Integridad Fallida"
          },
          "message": {
            "type": "string",
            "description": "Descripcion del error",
            "example": "No es posible enviar un POA vacío."
          },
          "detalles": {
            "description": "Detalles a solucionar para resolver el error",
            "example": [
              "El POA debe contener al menos 1 actividad registrada.",
              "Existen sub-actividades sin un auditor responsable asignado."
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "error",
          "message",
          "detalles"
        ]
      },
      "BancoActividadesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la actividad",
            "example": "banco-uuid-1"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de actividad",
            "example": "AUDITORIA"
          },
          "titulo": {
            "type": "string",
            "description": "Titulo de la actividad",
            "example": "Revision al rubro de Obra Publica"
          },
          "descripcion_corta": {
            "type": "string",
            "description": "Explicacion breve de la actividad",
            "example": "Verificar el adecuado cumplimiento de los procedimientos relativos a la..."
          }
        },
        "required": [
          "id",
          "tipo",
          "titulo",
          "descripcion_corta"
        ]
      },
      "BancoActividadesDataDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Lista de actividades",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BancoActividadesDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "BancoIdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del banco de actividades",
            "example": "banco-uuid-1"
          },
          "titulo": {
            "type": "string",
            "description": "Titulo de la actividad",
            "example": "Revisión al rubro de Obra Pública"
          },
          "justificacion_plantilla": {
            "type": "string",
            "description": "Justificación de la actividad",
            "example": "Verificar el adecuado cumplimiento a los procedimientos relativos a la planeación, programación..."
          },
          "objetivo_gen_plantilla": {
            "type": "string",
            "description": "Objetivo general de la actividad",
            "example": "Asegurar la correcta aplicación de los recursos destinados a la obra pública..."
          }
        },
        "required": [
          "id",
          "titulo",
          "justificacion_plantilla",
          "objetivo_gen_plantilla"
        ]
      },
      "ActividadSugeridaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la actividad sugerida",
            "example": "actividad-uuid-1"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripcion de la actividad sugerida",
            "example": "Elaborar acta circunstanciada para desincorporar del Patrimonio..."
          },
          "tipo_sugerido": {
            "type": "string",
            "description": "Que tipo de actividad es la sugerida",
            "example": "AUDITORIA"
          }
        },
        "required": [
          "id",
          "descripcion",
          "tipo_sugerido"
        ]
      },
      "ActividadSugeridaDataDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Lista de actividades sugeridas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActividadSugeridaDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}